src/core_pipelines/requirements.txt
--extra-index-url https://nexus-repo.mckinsey-solutions.com/nexus/repository/opm-pypi-releases/simple

adtk>=0.6.2
kedro[pandas.CSVDataSet, pandas.ExcelDataSet, pandas.ParquetDataSet]>=0.16.6
matplotlib>=3.0.3, <3.4.3
markdown~=3.3.0
m2r~=0.2
numexpr~=2.0
numpy>=1.18, <1.20.0
optimizer~=2.0
optimus_core>=0.1.3
pandas~=1.3.0
pandas-profiling[notebook, html]~=2.0
plotly>=4.14.3
PyPDF2>=1.26, <2.0
pymc3>=3.8
pytz>=2019.3
tune_sklearn>=0.2.0
scikit-learn>=0.21.3
seaborn>=0.10.1, <1.0.0
tqdm>=4.0, <5.0.0
xgboost>0.90, <2.0.0
